requirements.txt
Django==4.2
darker==2.1.1
django-environ==0.11.2
gunicorn==22.0.0
psycopg2-binary==2.9.9
djangorestframework==3.15.1
